Enum wafer::config::CentralDifference

pub enum CentralDifference {
    ThreePoint,
    FivePoint,
    SevenPoint,
}

Sets the precision of the central difference formalism.

Variants
ThreePoint

3 point, good to 𝓞(grid.dn²).

FivePoint

5 point, good to 𝓞(grid.dn⁴).

SevenPoint

7 point, good to 𝓞(grid.dn⁶).

Methods

impl CentralDifference

pub fn bb(&self) -> usize

Grabs the B ounding B ox size for the current precision.

pub fn ext(&self) -> usize

Grabs how much the work area is extended in one direction for the current precision.