Struct wafer::config::Output

pub struct Output {
    pub screen_update: u64,
    pub snap_update: Option<u64>,
    pub file_type: FileType,
    pub save_wavefns: bool,
    pub save_potential: bool,
}

Identifies the frequency of output to the screen or disk, as well as toggling the output of wavefunction and potential data.

Fields
screen_update: u64

How many steps should the system evolve before outputting information to the screen.

snap_update: Option<u64>

Optional: How many steps should the system evolve before saving a partially converged wavefunction.

file_type: FileType

File format to be used for output. Messagepack is the smallest (and fastest) option, but not human readable. Structured text options are json and yaml, then for a complete plain text option there is csv.

save_wavefns: bool

Should wavefunctions be saved at all? Not necessary if energy values are the only interest. Each excited state is saved once it is converged or if max_steps is reached.

save_potential: bool

Should the potential be saved for reference? This is output at the start of the simulation.