Module wafer::potential

Handles the potential generation, binding energy offsets, callouts to files or scripts if needed etc.

Re-exports

use ndarray::Array3;

use ndarray::Zip;

use ndarray_parallel::prelude::*;

use noisy_float::prelude::*;

use slog::Logger;

use std::f64::consts::PI;

use std::f64::MAX;

use config::Config;

use config::Grid;

use config::Index3;

use config::PotentialType;

use errors::*;

use grid;

use input;

use output;

Structs

PotentialSubSingle
A single value struct for potential_sub outputs that do not require an entire array.

Potentials
Holds the potential arrays for the current simulation.

Constants

_IMPL_DESERIALIZE_FOR_PotentialSubSingle

_IMPL_SERIALIZE_FOR_PotentialSubSingle

Functions

alphas
Running coupling. Used for Cornell potentials.

calculate_r2
Calculates squared distance

generate
A public wrapper around potential. Where potential does the calculation for a single point, generate builds the entire grid.

load_arrays
Handles the potential loading from file, or generating depending on configuration Ancillary arrays are also generated here.

mu
Debye screening mass. Used for Cornell potentials.

potential
Generates a potential for the current simulation at a particular index point.

potential_sub
Calculate binding energy offset (if any). Follows the potential input/output arguments. FullCornell, and subsequent potentials that require indexed values must call potential_sub_idx.

potential_sub_idx
Calculate binding energy offset (if any). Follows the potential input/output arguments. Used if calculation requires indexing. If not, call potential_sub instead. Currency only FullCornell requires this routine.